requirements.txt
agate==1.6.3
amqp==5.1.1
asgiref==3.5.0
asn1crypto==1.5.1
attrs==21.4.0
Babel==2.10.1
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.6
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.8
dbt-core==1.0.5
dbt-extractor==0.4.1
dbt-snowflake==1.0.1
defusedxml==0.7.1
Django==3.2.13
django-celery-beat==2.2.1
django-cors-headers==3.11.0
django-enumchoicefield==3.0.0
django-templated-mail==1.1.1
django-timezone-field==4.2.3
djangorestframework==3.13.1
#djangorestframework-simplejwt==5.1.0
djoser==2.1.0
gunicorn==20.0.4
future==0.18.2
gevent==21.12.0
greenlet==1.1.2
hologram==0.0.14
idna==3.3
importlib-metadata==4.11.3
isodate==0.6.1
itypes==1.2.0
Jinja2==2.11.3
jsonschema==3.1.1
keyring==23.5.0
kombu==5.2.4
leather==0.3.4
Logbook==1.5.3
MarkupSafe==2.0.1
mashumaro==2.9
minimal-snowplow-tracker==0.0.2
msgpack==1.0.3
networkx==2.8
oauthlib==3.2.0
oscrypto==1.3.0
packaging==21.3
parsedatetime==2.4
prompt-toolkit==3.0.29
pycparser==2.21
pycryptodomex==3.14.1
PyJWT==2.3.0
pyOpenSSL==21.0.0
pyparsing==3.0.8
pyrsistent==0.18.1
python-crontab==2.6.0
python-dateutil==2.8.2
python-dotenv==0.20.0
python-slugify==6.1.1
python3-openid==3.2.0
pytimeparse==1.1.8
pytz==2022.1
pywin32-ctypes==0.2.0
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
snowflake-connector-python==2.7.6
social-auth-app-django==4.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
text-unidecode==1.3
typing-extensions==3.10.0.2
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.1.1
zipp==3.8.0
zope.event==4.5.0
zope.interface==5.4.0
